started with strategy Execution Cancellation
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C30004

+++ BusinessProcess/Code/Vanrise.BusinessProcess.Client/BPClient.cs
@@ -123,7 +123,7 @@ namespace Vanrise.BusinessProcess.Client
             return dataManager.GetTrackingsFrom(input);
         }
 
-        public Dictionary<int, BPInstanceStatus> GetProcessesStatuses(List<int> Ids)
+        public Dictionary<long, BPInstanceStatus> GetProcessesStatuses(List<long> Ids)
         {
             return GetProcessesStatuses(Ids);
         }
